+++ R/ML_AdaBagModel.R
@@ -34,10 +34,11 @@
 #' @examples
 #' fit(Species ~ ., data = iris, model = AdaBagModel(mfinal = 5))
 #'
-AdaBagModel <- function(mfinal = 100, minsplit = 20,
-                        minbucket = round(minsplit/3), cp = 0.01,
-                        maxcompete = 4, maxsurrogate = 5, usesurrogate = 2,
-                        xval = 10, surrogatestyle = 0, maxdepth = 30) {
+AdaBagModel <- function(
+  mfinal = 100, minsplit = 20, minbucket = round(minsplit/3), cp = 0.01,
+  maxcompete = 4, maxsurrogate = 5, usesurrogate = 2, xval = 10,
+  surrogatestyle = 0, maxdepth = 30
+) {
 
   args <- params(environment())
   is_main <- names(args) %in% "mfinal"
@@ -51,7 +52,7 @@ AdaBagModel <- function(mfinal = 100, minsplit = 20,
     response_types = "factor",
     predictor_encoding = "terms",
     params = params,
-    grid = function(x, length, ...) {
+    grid = function(length, ...) {
       list(
         mfinal = round(seq_range(0, 25, c(1, 200), length + 1)),
         maxdepth = 1:min(length, 30)