+++ R/ML_BlackBoostModel.R
@@ -50,14 +50,13 @@
 #'
 #' fit(type ~ ., data = Pima.tr, model = BlackBoostModel)
 #'
-BlackBoostModel <- function(family = NULL, mstop = 100, nu = 0.1,
-                            risk = c("inbag", "oobag", "none"),
-                            stopintern = FALSE, trace = FALSE,
-                            teststat = c("quadratic", "maximum"),
-                            testtype = c("Teststatistic", "Univariate",
-                                         "Bonferroni", "MonteCarlo"),
-                            mincriterion = 0, minsplit = 10, minbucket = 4,
-                            maxdepth = 2, saveinfo = FALSE, ...) {
+BlackBoostModel <- function(
+  family = NULL, mstop = 100, nu = 0.1, risk = c("inbag", "oobag", "none"),
+  stopintern = FALSE, trace = FALSE, teststat = c("quadratic", "maximum"),
+  testtype = c("Teststatistic", "Univariate", "Bonferroni", "MonteCarlo"),
+  mincriterion = 0, minsplit = 10, minbucket = 4, maxdepth = 2,
+  saveinfo = FALSE, ...
+) {
 
   teststat <- match.arg(teststat)
   testtype <- match.arg(testtype)
@@ -79,7 +78,7 @@ BlackBoostModel <- function(family = NULL, mstop = 100, nu = 0.1,
                        "numeric", "PoissonVariate", "Surv"),
     predictor_encoding = "terms",
     params = params,
-    grid = function(x, length, ...) {
+    grid = function(length, ...) {
       list(
         mstop = round(seq_range(0, 50, c(1, 1000), length + 1)),
         maxdepth = 1:min(length, 10)
@@ -88,12 +87,13 @@ BlackBoostModel <- function(family = NULL, mstop = 100, nu = 0.1,
     fit = function(formula, data, weights, family = NULL, ...) {
       if (is.null(family)) {
         family <- switch_class(response(data),
-                               BinomialVariate = mboost::Binomial(type = "glm"),
-                               factor = mboost::Binomial(),
-                               NegBinomialVariate = mboost::NBinomial(),
-                               numeric = mboost::Gaussian(),
-                               PoissonVariate = mboost::Poisson(),
-                               Surv = mboost::CoxPH())
+          "BinomialVariate" = mboost::Binomial(type = "glm"),
+          "factor" = mboost::Binomial(),
+          "NegBinomialVariate" = mboost::NBinomial(),
+          "numeric" = mboost::Gaussian(),
+          "PoissonVariate" = mboost::Poisson(),
+          "Surv" = mboost::CoxPH()
+        )
       }
       mboost::blackboost(formula, data = as.data.frame(data),
                          na.action = na.pass, weights = weights,