+++ R/ML_EarthModel.R
@@ -43,10 +43,10 @@
 #' model_fit <- fit(Species ~ ., data = iris, model = EarthModel)
 #' varimp(model_fit, metric = "nsubsets", scale = FALSE)
 #'
-EarthModel <- function(pmethod = c("backward", "none", "exhaustive", "forward",
-                                   "seqrep", "cv"),
-                       trace = 0, degree = 1, nprune = NULL,
-                       nfold = 0, ncross = 1, stratify = TRUE) {
+EarthModel <- function(
+  pmethod = c("backward", "none", "exhaustive", "forward", "seqrep", "cv"),
+  trace = 0, degree = 1, nprune = NULL, nfold = 0, ncross = 1, stratify = TRUE
+) {
 
   pmethod <- match.arg(pmethod)
 
@@ -72,8 +72,8 @@ EarthModel <- function(pmethod = c("backward", "none", "exhaustive", "forward",
       ), name = "earth_exports")
 
       glm <- list(family = switch_class(response(data),
-                                        factor = "binomial",
-                                        numeric = "gaussian"))
+                                        "factor" = "binomial",
+                                        "numeric" = "gaussian"))
       eval_fit(data,
                formula = earth::earth(formula, data = as.data.frame(data),
                                       weights = weights, glm = glm, ...),